feat: reorganize app structure with separate API and models directories
- Create app/api/ directory for API endpoints
  - Split endpoints into core.py (basic endpoints) and scraping.py (scraping endpoints)
  - Move FastAPI app creation to api/main.py
- Create app/models/ directory for Pydantic models
  - Split each model class into individual files
  - ScrapingRequest, ScrapingResponse, ExecutionHistoryItem, HistoryResponse, StatusResponse
- Update import statements across the application
- Maintain backward compatibility with existing functionality
- All tests passing with new structure

This reorganization improves code maintainability and follows better separation of concerns principles.
