+++ app/__init__.py
@@ -1,0 +1,5 @@
+"""Application package."""
+
+from .main import app
+
+__all__ = ["app"]
