+++ app/api/__init__.py
@@ -1,0 +1,5 @@
+"""API package for the application."""
+
+from .main import app
+
+__all__ = ["app"]
