+++ run.py
@@ -4,11 +4,11 @@ Startup script for the 6MLET Tech Challenge Delivery 01 application.
 """
 
 import uvicorn
-from app.main import app
+from app import app
 
 if __name__ == "__main__":
     uvicorn.run(
-        "app.main:app",
+        "app.api.main:app",
         host="0.0.0.0",
         port=8000,
         reload=True,
