+++ tests/test_main.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from app.main import app
+from app import app
 
 client = TestClient(app)
